Modified PIO script to use closest possible baud-rate
Any timing which is generated in the PIO must be a divider of 125MHz (8ns resolution). Therefore, some baud-rates are unachievable.

The script has been modified to use the closest possible baudrate to the one specified, which can be generated without timing-error.

For example: configuring a baudrate of 100050 (not possible without error at 125MHz/8ns resolution) will print a warning and then change the baudrate to 100080 (no error at 125MHz/8ns resolution).